refactor: extract font plugins into src/plugins/fonts.ts
Move all Google Fonts and local fonts plugin logic out of the monolithic
index.ts into a single focused file (src/plugins/fonts.ts). The new file
exports createGoogleFontsPlugin() and createLocalFontsPlugin() factory
functions, plus the shared helpers VIRTUAL_GOOGLE_FONTS,
RESOLVED_VIRTUAL_GOOGLE_FONTS, generateGoogleFontsVirtualModule, and
parseStaticObjectLiteral (which index.ts also uses for tsconfig parsing).

index.ts now imports these and delegates to the factories, removing ~340
lines of inline plugin code from the main plugin file.
